+++ src/main/java/com/nami/webstore/model/Pagamento.java
@@ -4,6 +4,7 @@ import com.nami.webstore.enums.MetodoPagamento;
 import com.nami.webstore.enums.StatusPagamento;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "pagamentos")
@@ -24,7 +25,7 @@ public class Pagamento {
     private StatusPagamento status = StatusPagamento.PENDENTE;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private double valor;
+    private BigDecimal valor;
 
     @Column(name = "id_transacao", length = 100)
     private String idTransacao;
@@ -40,8 +41,8 @@ public class Pagamento {
     public void setMetodo(MetodoPagamento metodo) { this.metodo = metodo; }
     public StatusPagamento getStatus() { return status; }
     public void setStatus(StatusPagamento status) { this.status = status; }
-    public double getValor() { return valor; }
-    public void setValor(double valor) { this.valor = valor; }
+    public BigDecimal getValor() { return valor; }
+    public void setValor(BigDecimal valor) { this.valor = valor; }
     public String getIdTransacao() { return idTransacao; }
     public void setIdTransacao(String idTransacao) { this.idTransacao = idTransacao; }
     public LocalDateTime getPagoEm() { return pagoEm; }
